+++ src/Modules/Workflows/Domain/Steps/Triggers/Definitions/OnPostRowAction.php
@@ -1,0 +1,175 @@
+<?php
+
+namespace PublishPress\Future\Modules\Workflows\Domain\Steps\Triggers\Definitions;
+
+use PublishPress\Future\Modules\Workflows\Interfaces\StepTypeInterface;
+use PublishPress\Future\Modules\Workflows\Models\StepTypesModel;
+
+class OnPostRowAction implements StepTypeInterface
+{
+    public static function getNodeTypeName(): string
+    {
+        return "trigger/core.post-row-action";
+    }
+
+    public function getElementaryType(): string
+    {
+        return StepTypesModel::STEP_TYPE_TRIGGER;
+    }
+
+    public function getReactFlowNodeType(): string
+    {
+        return "trigger";
+    }
+
+    public function getBaseSlug(): string
+    {
+        return "onPostRowAction";
+    }
+
+    public function getLabel(): string
+    {
+        return __("Post row action", "post-expirator");
+    }
+
+    public function getDescription(): string
+    {
+        return __(
+            "This trigger activates whenever a post is saved, regardless of whether it's newly created, imported, or updated.", // phpcs:ignore Generic.Files.LineLength.TooLong
+            "post-expirator"
+        );
+    }
+
+    public function getIcon(): string
+    {
+        return "media-document";
+    }
+
+    public function getFrecency(): int
+    {
+        return 1;
+    }
+
+    public function getVersion(): int
+    {
+        return 2;
+    }
+
+    public function getCategory(): string
+    {
+        return "post";
+    }
+
+    public function getSettingsSchema(): array
+    {
+        return [
+            [
+                "label" => __("Post Query", "post-expirator"),
+                "description" => __(
+                    "Specify the criteria for posts that will trigger this action.",
+                    "post-expirator"
+                ),
+                "fields" => [
+                    [
+                        "name" => "postQuery",
+                        "type" => "postQuery",
+                        "label" => __("Post query", "post-expirator"),
+                        "description" => __(
+                            "The query defines the posts that will trigger this action.",
+                            "post-expirator"
+                        ),
+                        "settings" => [
+                            "acceptsInput" => false,
+                            "isPostTypeRequired" => true,
+                            "postTypeDescription" => __("Select the post types that will trigger this action.", "post-expirator"),
+                            "postIdDescription" => __("Enter one or more post IDs. Leave empty to include all posts.", "post-expirator"),
+                            "postStatusDescription" => __("If selected, only posts with these statuses will trigger this action.", "post-expirator"),
+                        ],
+                        "default" => [
+                            "postSource" => "custom",
+                            "postType" => ["post"],
+                            "postId" => [],
+                            "postStatus" => [],
+                        ],
+                    ],
+                    [
+                        "name" => "actionLabel",
+                        "type" => "text",
+                        "label" => __("Action label", "post-expirator"),
+                        "description" => __(
+                            "The label of the action to perform when the post row is clicked.",
+                            "post-expirator"
+                        ),
+                        "settings" => [
+                            "placeholder" => __("Run workflow", "post-expirator"),
+                        ],
+                    ]
+                ]
+            ]
+        ];
+    }
+
+    public function getValidationSchema(): array
+    {
+        return [
+            "settings" => [
+                "rules" => [
+                    [
+                        "rule" => "required",
+                        "field" => "postQuery.postType",
+                        "label" => __("Post Type", "post-expirator"),
+                    ],
+                    [
+                        "rule" => "dataType",
+                        "field" => "postQuery.postId",
+                        "type" => "integerList",
+                        "label" => __("Post ID", "post-expirator"),
+                    ],
+                ],
+            ],
+            "connections" => [
+                "rules" => [
+                    [
+                        "rule" => "hasOutgoingConnection",
+                    ],
+                ]
+            ]
+        ];
+    }
+
+    public function getOutputSchema(): array
+    {
+        return [
+            [
+                'name' => 'post',
+                'type' => 'post',
+                'label' => __("Triggered post", "post-expirator"),
+                'description' => __("The post where the action was triggered.", "post-expirator"),
+            ]
+        ];
+    }
+
+    public function getCSSClass(): string
+    {
+        return "react-flow__node-genericTrigger";
+    }
+
+    public function getHandleSchema(): array
+    {
+        return [
+            "target" => [],
+            "source" => [
+                [
+                    "id" => "output",
+                    "left" => "50%",
+                    "label" => __("Next", "post-expirator"),
+                ]
+            ]
+        ];
+    }
+
+    public function isProFeature(): bool
+    {
+        return true;
+    }
+}